Use Udacity final project version
This commit updates the source code to that given by Udacity as
the starting point for the final set of updates to the web app.
I would rather work on that and have more in common with the
videos than continue with mine and potentially miss out on
styling that I have not yet encountered.